pyproject.toml
[tool.poetry]
name = "grid-wit"
version = "0.1.0"
description = "A crossword puzzle generator using historical NYT crossword data."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11"
pandas = "^1.3.0"
numpy = "^1.21.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
requests = "^2.26.0"
puzpy = "^0.2.0"
flask = "^3.0.3"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.0"
sqlalchemy-utils = "^0.41.1"
gunicorn = "^21.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
